backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
pydantic==2.3.0
python-multipart==0.0.6
cloudinary==1.34.0
google-cloud-vision==3.4.5
requests==2.31.0
fastapi-throttling==0.1.0
python-dotenv==1.0.0 